First node test
