+++ derive_helpers/Cargo.toml
@@ -19,6 +19,7 @@ strum = { workspace = true }
 strum_macros = { workspace = true }
 thiserror = { workspace = true }
 paste = { workspace = true }
+insta = { workspace = true }
 
 
 serde = { workspace = true }
